+++ packages/server/src/api/muster/muster.compliance.spec.ts
@@ -1,11 +1,11 @@
 import { expect } from 'chai';
 import { toMusterCompliance } from './muster.compliance';
-import { getRosterTestDataSimple } from '../../util/test-utils/roster.test-generator';
+import { rosterTestDataSimple } from '../../util/test-utils/data/roster-generator';
 
 describe('Muster Compliance', () => {
   describe('toMusterCompliance()', () => {
     it('should calculate muster compliance', () => {
-      const rosterIndividualsInformation = getRosterTestDataSimple(10, 2, 1);
+      const rosterIndividualsInformation = rosterTestDataSimple(10, 2, 1);
       // let recordedObservations;
       // let musteringOpportunities;
       // const musterCompliance = toMusterCompliance(rosterIndividualsInformation, recordedObservations, musteringOpportunities);
