+++ packages/server/src/util/test-utils/data/report-schema-generator.ts
@@ -1,7 +1,7 @@
-import { Org } from '../../api/org/org.model';
-import { defaultReportSchemas, ReportSchema } from '../../api/report-schema/report-schema.model';
+import { Org } from '../../../api/org/org.model';
+import { defaultReportSchemas, ReportSchema } from '../../../api/report-schema/report-schema.model';
 
-export function getReportSchemaTestData(org: Org) {
+export function reportSchemaTestData(org: Org) {
   let reportSchemas = ReportSchema.create(defaultReportSchemas);
   reportSchemas = reportSchemas.concat(ReportSchema.create(defaultReportSchemas));
   for (let i = 0; i < reportSchemas.length; i++) {
